requirements.txt
flask>=2.0.1
google-generativeai
